AIforSBDv1.0/data_ingestion/requirements.txt
# Requirements for data ingestion module (isolated)
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
azure-search-documents
azure-devops
aiohttp==3.9.1
pydantic==2.5.3
python-dotenv==1.0.0
azure-monitor-opentelemetry==1.2.0
beautifulsoup4==4.12.2
markdownify==0.11.6
tenacity==8.2.3
